Move from c++17 to c++20, Make visuals better, Fix a bug in Engine panel that cruses when no engine found, Remove ECO example code, Add a way to let user choose whether to be asked for new variation, Add a way to let user choose whether to see Pgn files only, Start changing the name to Chess Lab, Move Engine Downloadable files to DrobBox and make the nacessary changes, Fix a bug in formated Fen, Fix a bug when you go back from an an pan san move a piece disapears, Fix a bug that checks are being written in the next move, Make GameManager a bit safer

+++ LightSource-Lobby/src/windowsMain.h
@@ -33,17 +33,17 @@ public:
         // if (ChildStd_OUT_Rd)
         //     m_ChildStd_OUT_Rd = ChildStd_OUT_Rd;
         if (!CreatePipe(&m_ChildStd_OUT_Rd, &ChildStd_OUT_Wr, &saAttr, 0))
-            throw std::runtime_error("StdoutRd CreatePipe");
+            printf("StdoutRd CreatePipe");
 
         if (!SetHandleInformation(m_ChildStd_OUT_Rd, HANDLE_FLAG_INHERIT, 0))
-            throw std::runtime_error("Stdout SetHandleInformation");
+            printf("Stdout SetHandleInformation");
 
 
         if (!CreatePipe(&ChildStd_IN_Rd, &m_ChildStd_IN_Wr, &saAttr, 0))
-            throw std::runtime_error("Stdin CreatePipe");
+            printf("Stdin CreatePipe");
 
         if (!SetHandleInformation(m_ChildStd_IN_Wr, HANDLE_FLAG_INHERIT, 0))
-            throw std::runtime_error("Stdin SetHandleInformation");
+            printf("Stdin SetHandleInformation");
 
         TCHAR szCmdline[] = L"";
         PROCESS_INFORMATION piProcInfo;
@@ -75,7 +75,7 @@ public:
             &piProcInfo);  // receives PROCESS_INFORMATION 
 
         if (!bSuccess)
-            throw std::runtime_error("CreateProcess");
+            printf("CreateProcess");
         else
         {
             CloseHandle(ChildStd_OUT_Wr);
@@ -109,14 +109,15 @@ public:
         CloseHandle(m_ChildStd_IN_Wr);
         CloseHandle(m_ChildStd_OUT_Rd);
 
-        TerminateProcess(m_Process, 0);
+        if (IsProcessActive())
+            TerminateProcess(m_Process, 0);
 
         CloseHandle(m_Process);
     }
 
     ~Process()
     {
-       
+        //EndProcess();
     }
 
     void Write(std::string message)
@@ -128,7 +129,7 @@ public:
 
         bSuccess = WriteFile(m_ChildStd_IN_Wr, message.c_str(), message.size(), &dwWritten, NULL);
         if (!bSuccess)
-            throw std::runtime_error("Write to Process");
+            printf("Write to Process");
     }
 
     std::string Read()
@@ -145,7 +146,7 @@ public:
             if (!bSuccess || dwRead == 0)
                 break;
             if (!bSuccess)//will not be hitted 
-                throw std::runtime_error("Read Process");
+                printf("Read Process");
             output += std::string(chBuf, dwRead);
         } while (dwRead == BUFSIZE);
 
@@ -153,56 +154,10 @@ public:
     }
 
 private:
-   
+
     HANDLE m_ChildStd_IN_Wr = NULL;
     HANDLE m_ChildStd_OUT_Rd = NULL;
     HANDLE m_Process = NULL;
 
 #endif // PLATFORM_WINDOWS_64
 };
-
-#if 0
-int notmain()
-{
-    std::string overall;
-    float score = 0;
-
-    Process stockfish(L"stockfish.exe", L"");
-    stockfish.Write("ucinewgame");
-    stockfish.Write("position fen 4r3/6pp/pp1k1b2/2p1N3/5P2/2P3P1/P4K1P/4R3 w - - 5 31");
-    stockfish.Write("d");
-    using namespace std::chrono_literals;
-    std::this_thread::sleep_for(1s);
-    std::cout << stockfish.Read() << '\n' << score;
-    stockfish.Write("go infinite");
-    for (int i = 0; i < 200; i++)
-    {
-        overall.clear();
-        using namespace std::chrono_literals;
-        std::this_thread::sleep_for(0.4s);
-        overall += stockfish.Read();
-        int index = overall.find("score cp");
-        if (index + 9 < overall.size() && index >= 0)
-        {
-            index += 9;
-            std::string strscore;
-            for (int j = index; j < overall.size(); j++)
-            {
-                if (overall[j] == ' ')
-                    break;
-                strscore += overall[j];
-            }
-            score = (float)std::stoi(strscore) / 100.0f;
-        }
-        std::cout << '\r' << score;
-    }
-
-    //stockfish.Write("stop");
-    using namespace std::chrono_literals;
-    std::this_thread::sleep_for(0.4s);
-    std::cout << stockfish.Read();
-
-    
-    printf("\n->End of parent execution.\n");
-}
-#endif
